+++ packages/core/src/compiler/scanner/subscanners/index.ts
@@ -14,4 +14,8 @@ export * from "./FormRequestScanner";
 export * from "./ModelScanner";
 export * from "./RouteScanner";
 export * from "./InvalidationResolver";
+export * from "./typeDeriverUtils";
+export * from "./ValidationRuleFieldLowerer";
+export * from "./RequestTypeDeriver";
+export * from "./SemanticTypeDeriver";
 export * from "./TypeDeriver";